Resouces states managing + temporary remove imgui

+++ QuakeRenderEngine/source/RendererPasses/ModelsRenderPass/ModelsPassProvider.cpp
@@ -40,7 +40,7 @@ const D3D11_INPUT_ELEMENT_DESC  SingleFrameInputElements[] =
     { "TEXCOORD",           0, DXGI_FORMAT_R32G32_FLOAT,       0, D3D11_APPEND_ALIGNED_ELEMENT, D3D11_INPUT_PER_VERTEX_DATA, 0 },
 };
 
-Renderer::InputLayoutDescription ModelsPassProvider::GetInputLayoutDescription(size_t definesFlags)
+Renderer::InputLayoutDescription ModelsPassProvider::GetInputLayoutDescription(uint32_t definesFlags)
 {
     if (definesFlags & ModelDefines::MLERP) {
         return Renderer::InputLayoutDescription{ (void*)LerpInputElements, std::size(LerpInputElements) };
