+++ QuakeRenderEngine/source/RendererPasses/SSAORenderPass/SSAORenderPass.h
@@ -21,8 +21,8 @@ class SSAORenderPass:public BaseRenderPass {
 
 	struct SSAORenderPassProvider : public Renderer::IStateProvider {
 		SSAORenderPassProvider();
-		virtual void PatchPipelineState(Renderer::Pipeline* refToPS, size_t definesFlags) override;
-		virtual  Renderer::InputLayoutDescription GetInputLayoutDescription(size_t definesFlags) override;
+		virtual void PatchPipelineState(Renderer::Pipeline* refToPS, uint32_t definesFlags) override;
+		virtual  Renderer::InputLayoutDescription GetInputLayoutDescription(uint32_t definesFlags) override;
 		virtual const char* GetShaderName() override;
 		Renderer::PipelineFactoryDescription GetFactoryDescription() override;
 	};
