+++ QuakeRenderEngine/source/RendererPasses/UIRenderPass/UIRenderPassProvider.cpp
@@ -28,7 +28,7 @@ const D3D11_INPUT_ELEMENT_DESC UIInputElements[] =
     {"TEXCOORD",0,DXGI_FORMAT_R32G32_FLOAT,0, D3D11_APPEND_ALIGNED_ELEMENT,D3D11_INPUT_PER_VERTEX_DATA,0},
 };
 
-Renderer::InputLayoutDescription UIRenderPassProvider::GetInputLayoutDescription(size_t definesFlags)
+Renderer::InputLayoutDescription UIRenderPassProvider::GetInputLayoutDescription(uint32_t definesFlags)
 {
     return Renderer::InputLayoutDescription{(void*)UIInputElements,std::size(UIInputElements)};
 }
